+++ src/spawn/__init__.py
@@ -3,4 +3,4 @@ from importlib.metadata import version, PackageNotFoundError
 try:
     __version__ = version("spawn")
 except PackageNotFoundError:
-    __version__ = "1.0.3"
+    __version__ = "1.0.4"